# docs/swagger.yaml
basePath: /api/v1
host: 127.0.0.1:8080
info:
  contact:
    email: xx@xxx.xx
    name: Villero
    url: https://xxx.xx
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://xxx.xx
  title: Swagger Example API
  version: 1.6.5
paths:
  /api/v1/articles:
    get:
      parameters:
      - description: TagID
        in: body
        name: tag_id
        schema:
          type: integer
      - description: State
        in: body
        name: state
        schema:
          type: integer
      - description: CreatedBy
        in: body
        name: created_by
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 获取多篇文章
    post:
      parameters:
      - description: TagID
        in: body
        name: tag_id
        required: true
        schema:
          type: integer
      - description: Title
        in: body
        name: title
        required: true
        schema:
          type: string
      - description: Desc
        in: body
        name: desc
        required: true
        schema:
          type: string
      - description: Content
        in: body
        name: content
        required: true
        schema:
          type: string
      - description: CreatedBy
        in: body
        name: created_by
        required: true
        schema:
          type: string
      - description: State
        in: body
        name: state
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 新增文章
  /api/v1/articles/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 删除文章
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 获取单个文章
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: TagID
        in: body
        name: tag_id
        schema:
          type: string
      - description: Title
        in: body
        name: title
        schema:
          type: string
      - description: Desc
        in: body
        name: desc
        schema:
          type: string
      - description: Content
        in: body
        name: content
        schema:
          type: string
      - description: ModifiedBy
        in: body
        name: modified_by
        required: true
        schema:
          type: string
      - description: State
        in: body
        name: state
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 修改文章
  /api/v1/tags:
    get:
      parameters:
      - description: Name
        in: query
        name: name
        type: string
      - description: State
        in: query
        name: state
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 获得多个文章标签
    post:
      parameters:
      - description: Name
        in: query
        name: name
        required: true
        type: string
      - description: State
        in: query
        name: state
        type: integer
      - description: CreatedBy
        in: query
        name: created_by
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 新增文章标签
  /api/v1/tags/{id}:
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Name
        in: body
        name: name
        required: true
        schema:
          type: string
      - description: State
        in: body
        name: state
        schema:
          type: integer
      - description: ModifiedBy
        in: body
        name: modified_by
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 删除文章标签
  /auth:
    get:
      parameters:
      - description: userName
        in: query
        name: username
        required: true
        type: string
      - description: password
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 获取token
swagger: "2.0"